pyproject.toml
[project]
name = "dog-breed-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black==24.8.0",
    "kaggle==1.6.17",
    "lightning==2.4.0",
    "loguru==0.7.2",
    "pytorch-lightning==2.4.0",
    "rich==13.8.1",
    "tensorboard==2.17.1",
    "timm==1.0.9"
]